requirements.txt
requests
numpy
pandas
matplotlib
seaborn
scikit-learn
pydot
graphviz
plotly
nbformat>=4.2.0
pyyaml